Update README.md
This repository contains a code for visualization of the registered landslides in Colombia from the Global Landslide Catalog, provided by NASA's Open Data Portal. The results displayed the distribution of registered landslides in Bogotá city. 
With the landslide hazard maps available in the open data source by the Bogotá's Institute for Risk Management and Climate Change (IDIGER) and the characterization of the exposed elements that I conducted in the study: "Landslide risk analysis and assessment in the Rafael Uribe Uribe locality, Bogotá, Colombia." The script executes a physical landslide risk assessment in the 1.024 exposed elements identified in the Rafael Uribe Uribe locality. Additionally, a statistical analysis shows the distribution of the exposed elements in high risk in terms of potential losses.
Lastly, the database with the characterization of the exposed elements is attached. Also, the shapefile I used to execute the code and the list of abbreviations from the database (adapted from the original language into English).